バケツリレー問題の解決
