+++ khaki_player/lib/components/main.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:khaki_player/components/khakiPlayerPage.dart';
+import 'package:khaki_player/models/adioPlayerModel.dart';
+import 'package:provider/provider.dart';
 
 void main() {
   runApp(MyApp());
@@ -10,24 +12,16 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Flutter Demo',
       theme: ThemeData(
-        // This is the theme of your application.
-        //
-        // Try running your application with "flutter run". You'll see the
-        // application has a blue toolbar. Then, without quitting the app, try
-        // changing the primarySwatch below to Colors.green and then invoke
-        // "hot reload" (press "r" in the console where you ran "flutter run",
-        // or simply save your changes to "hot reload" in a Flutter IDE).
-        // Notice that the counter didn't reset back to zero; the application
-        // is not restarted.
         primarySwatch: Colors.cyan,
-        // This makes the visual density adapt to the platform that you run
-        // the app on. For desktop platforms, the controls will be smaller and
-        // closer together (more dense) than on mobile platforms.
         visualDensity: VisualDensity.adaptivePlatformDensity,
       ),
-      home: KhakiPlayerPage()
+      home: MultiProvider(
+        providers: [
+          ChangeNotifierProvider(create: (_) => AudioPlayerModel())
+        ],
+        child: KhakiPlayerPage(),
+      )
     );
   }
 }
